refactor: Rename utils package to inspections
