+++ src/main/java/io/camunda/zeebe/bpmnassert/inspections/InspectionUtility.java
@@ -1,4 +1,4 @@
-package io.camunda.zeebe.bpmnassert.utils;
+package io.camunda.zeebe.bpmnassert.inspections;
 
 import io.camunda.zeebe.bpmnassert.filters.StreamFilter;
